Fix #711 and #714.

+++ core/src/main/java/org/jobrunr/utils/mapper/jsonb/NullSafeJsonBuilder.java
@@ -74,7 +74,7 @@ public class NullSafeJsonBuilder implements JsonObjectBuilder {
     }
 
     public NullSafeJsonBuilder add(String name, Integer value) {
-        if (value != null) add(name, value);
+        if (value != null) add(name, value.intValue());
         else addNull(name);
         return this;
     }
